+++ src/client/src/pages/settings/admin/userAdministration.tsx
@@ -21,10 +21,9 @@ import { useDeleteUserPrompts } from "./useDeleteUserPrompts.tsx";
 import { UserAdministrationContext } from "./userAdministrationContext.tsx";
 import { useSharedTableColumns } from "./useSharedTableColumns.tsx";
 
-export const UserTable: FC = () => {
+export const UserAdministration: FC = () => {
   const { t, i18n } = useTranslation();
   const [filterModel, setFilterModel] = useState<GridFilterModel>();
-  const [isLoading, setIsLoading] = useState(true);
   const history = useHistory();
   const { callApiWithErrorHandling, callApiWithRollback } = useApiRequest();
   const { statusColumn, getDeleteColumn } = useSharedTableColumns();
@@ -34,11 +33,9 @@ export const UserTable: FC = () => {
   const handleFilterModelChange = useCallback((newModel: GridFilterModel) => setFilterModel(newModel), []);
 
   useEffect(() => {
-    setIsLoading(true);
     const getUsers = async () => {
       const users: User[] = await callApiWithErrorHandling(fetchUsers, []);
       setUsers(users);
-      setIsLoading(false);
     };
     getUsers();
     setSelectedUser(null);
@@ -173,7 +170,7 @@ export const UserTable: FC = () => {
       getRowClassName={getRowClassName}
       rowHeight={44}
       sortingOrder={["asc", "desc"]}
-      loading={isLoading}
+      loading={!users?.length}
       onRowClick={handleRowClick}
       rowCount={users?.length}
       rows={users}