+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added a table displaying all the users in the admin settings.
 - Added a detail view for users in the admin settings.
+- Added a read-only table displaying all the workgroups in the admin settings.
 
 ### Changed